Tags are now joined to Tastes. Ugly though

+++ server/src/main/scala/com/panuleppaniemi/tasty/models/Tag.scala
@@ -2,10 +2,10 @@ package com.panuleppaniemi.tasty.models
 
 import scala.slick.driver.MySQLDriver.simple._
 
-case class Tag(id: Option[Long] = None, name: String)
+case class Tag(id: Option[Int] = None, name: String)
 
 object Tags extends Table[Tag]("tags") {
-  def id = column[Long]("id", O.PrimaryKey, O.AutoInc)
+  def id = column[Int]("id", O.PrimaryKey, O.AutoInc)
   def name = column[String]("name")
 
   def * = id.? ~ name <> (Tag, Tag.unapply(_))
